Creacion de registro y validar token para usuarios
